+++ CHANGELOG.md
@@ -10,9 +10,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 - `IDriver.bulkTransaction` to replace `DeltaGraph.apply` and make database specific bulk changes.
-- `EVAL_TYPE` links for `MethodReturn` and `BlockVertex` on the method stubs.
+- `IdentityStmt` is now handled as part of the `LOCAL` and `IDENTIFIER` cycle
+- `IdentityRef` is now handled under `projectOp`
 
 ### Fixed
+
+- External method <-`AST`- External type is now fixed.
+- `EVAL_TYPE` links for `MethodReturn` and `BlockVertex` on the method stubs.
 - `CacheOptions.cacheSize` is mutable via setters now
 
 ### Changed